refactor/#31: 파일 경로 최신화
